requirements.txt
Flask==3.0.0
requests==2.28.2
google-cloud-tasks==2.15.0
google-cloud-datastore==2.19.0
google-cloud-storage==2.18.0
google-cloud-storage-control==1.0.2
google-cloud-ndb==2.3.1

python-dotenv==1.0.1
Flask-CORS==5.0.0
